refactor(flight-tune): split the retry projection to meet the structure caps
The attempt projection, its cardinality rules, and the independent replay's
retry and score derivations each move to their own module, and the event
walk in both projections dispatches to named recorders instead of carrying
every arm inline.
